+++ packages/component-library/src/components/CollectionAccordion/CollectionAccordion.mock.ts
@@ -1,0 +1,28 @@
+import mockAccordion from '../Accordion/Accordion.mock';
+import mockTheme from '../../theme/mock.theme';
+
+export default {
+  __typename: 'Collection',
+  variant: 'accordion',
+  itemSpacing: 2,
+  items: [
+    {
+      ...mockAccordion,
+      variant: 'accordion-standard',
+      title: 'Stress and anxiety: What’s the difference?'
+    },
+    {
+      ...mockAccordion,
+      variant: 'accordion-standard',
+      title: 'What is depression? How do I know if I have it'
+    },
+    { ...mockAccordion, variant: 'accordion-standard', title: 'What is psychosis? How do I know if I have it?' },
+    {
+      ...mockAccordion,
+      variant: 'accordion-standard',
+      title: 'How soon should I seek help? Will this get better on its own?'
+    }
+  ],
+  itemsVariant: 'accordion-standard',
+  theme: [mockTheme]
+};