+++ packages/component-library/src/components/CollectionAccordion/index.ts
@@ -1,0 +1,1 @@
+export { default, CollectionAccordionProps } from './CollectionAccordion';