+++ packages/component-library/src/theme/createCollectionVariants.ts
@@ -45,10 +45,14 @@ export const collectionTilesVariant = (_: Theme) => ({
     variant: 'collection-tiles'
   },
   style: {
+    'display': 'flex',
+    'flexWrap': 'wrap',
+    'justifyContent': 'space-around',
+    'width': '100%',
+    'height': '100%',
     '& [class*="Section-root"] > [class*="Section-gridContainer"]': {
       '& > [class*="Section-gridItem"]': {
-        maxWidth: 1296 / 3,
-        marginBottom: 0
+        margin: 0
       }
     }
   }